requirements.txt
Flask
requests
python-dotenv
groq
PyQt5